Fix forge linting errors using forge formatter

+++ solidity/test/CSMM.t.sol
@@ -45,13 +45,7 @@ contract CSMMTest is Test, Deployers {
         address underlyingAsset,
         uint256 base_vts
     ) internal returns (Currency currency) {
-        IToken token = new IToken(
-            name,
-            symbol,
-            underlyingAsset,
-            address(rfs),
-            base_vts
-        );
+        IToken token = new IToken(name, symbol, underlyingAsset, address(rfs), base_vts);
 
         address[10] memory toApprove = [
             address(swapRouter),
@@ -71,10 +65,7 @@ contract CSMMTest is Test, Deployers {
         }
 
         // ! Make sure to approve the ITokens to take out 'underlyingAsset'
-        IERC20Minimal(underlyingAsset).approve(
-            address(token),
-            Constants.MAX_UINT256
-        );
+        IERC20Minimal(underlyingAsset).approve(address(token), Constants.MAX_UINT256);
         return Currency.wrap(address(token));
     }
 
@@ -87,49 +78,30 @@ contract CSMMTest is Test, Deployers {
         Currency _currencyB = deployMintAndApproveCurrency();
 
         Currency _currencyC = deployAndApproveITokens(
-            "Intents TOKEN0 Settlement Receipt",
-            "iTOKEN0R",
-            Currency.unwrap(_currencyA),
-            baseVts
+            "Intents TOKEN0 Settlement Receipt", "iTOKEN0R", Currency.unwrap(_currencyA), baseVts
         );
         Currency _currencyD = deployAndApproveITokens(
-            "Intents TOKEN1 Settlement Receipt",
-            "ITOKEN1R",
-            Currency.unwrap(_currencyB),
-            baseVts
+            "Intents TOKEN1 Settlement Receipt", "ITOKEN1R", Currency.unwrap(_currencyB), baseVts
         );
 
-        (_currency0, _currency1) = SortTokens.sort(
-            MockERC20(Currency.unwrap(_currencyA)),
-            MockERC20(Currency.unwrap(_currencyB))
-        );
+        (_currency0, _currency1) =
+            SortTokens.sort(MockERC20(Currency.unwrap(_currencyA)), MockERC20(Currency.unwrap(_currencyB)));
 
-        (_currency2, _currency3) = SortTokens.sort(
-            MockERC20(Currency.unwrap(_currencyC)),
-            MockERC20(Currency.unwrap(_currencyD))
-        );
+        (_currency2, _currency3) =
+            SortTokens.sort(MockERC20(Currency.unwrap(_currencyC)), MockERC20(Currency.unwrap(_currencyD)));
     }
 
     function deployCorePool() public {
         //  Deploy the pool without the hook
-        (corePoolKey, ) = initPool(
-            _currency2,
-            _currency3,
-            IHooks(address(0)),
-            3000,
-            SQRT_PRICE_1_1,
-            ZERO_BYTES
-        );
+        (corePoolKey,) = initPool(_currency2, _currency3, IHooks(address(0)), 3000, SQRT_PRICE_1_1, ZERO_BYTES);
     }
 
     function deployProxyPool() public {
         // Proxy pool needs hook not core pool
         address hookAddress = address(
             uint160(
-                Hooks.BEFORE_ADD_LIQUIDITY_FLAG |
-                    Hooks.BEFORE_SWAP_FLAG |
-                    Hooks.BEFORE_INITIALIZE_FLAG |
-                    Hooks.BEFORE_SWAP_RETURNS_DELTA_FLAG
+                Hooks.BEFORE_ADD_LIQUIDITY_FLAG | Hooks.BEFORE_SWAP_FLAG | Hooks.BEFORE_INITIALIZE_FLAG
+                    | Hooks.BEFORE_SWAP_RETURNS_DELTA_FLAG
             )
         );
 
@@ -138,14 +110,7 @@ contract CSMMTest is Test, Deployers {
         hook = CSMM(hookAddress);
 
         //  Deploy the pool with the hook
-        (proxyPoolKey, ) = initPool(
-            _currency0,
-            _currency1,
-            hook,
-            3000,
-            SQRT_PRICE_1_1,
-            ZERO_BYTES
-        );
+        (proxyPoolKey,) = initPool(_currency0, _currency1, hook, 3000, SQRT_PRICE_1_1, ZERO_BYTES);
     }
 
     function mintAndApproveHookToMintITokens() public {
@@ -170,7 +135,7 @@ contract CSMMTest is Test, Deployers {
         itoken1.wrap(custodian, mintAmount);
     }
 
- function setUp() public {
+    function setUp() public {
         deployFreshManagerAndRouters();
         deployCurrencies();
         deployCorePool();
@@ -195,12 +160,7 @@ contract CSMMTest is Test, Deployers {
     function test_canModifyLiquidityOfCoreHook() public {
         modifyLiquidityRouter.modifyLiquidity(
             corePoolKey,
-            IPoolManager.ModifyLiquidityParams({
-                tickLower: -60,
-                tickUpper: 60,
-                liquidityDelta: 1e18,
-                salt: bytes32(0)
-            }),
+            IPoolManager.ModifyLiquidityParams({tickLower: -60, tickUpper: 60, liquidityDelta: 1e18, salt: bytes32(0)}),
             ZERO_BYTES
         );
     }
@@ -217,12 +177,10 @@ contract CSMMTest is Test, Deployers {
             ZERO_BYTES
         );
 
-        PoolSwapTest.TestSettings memory settings = PoolSwapTest.TestSettings({
-            takeClaims: false,
-            settleUsingBurn: false
-        });
+        PoolSwapTest.TestSettings memory settings =
+            PoolSwapTest.TestSettings({takeClaims: false, settleUsingBurn: false});
 
-        uint selfBalanceOfTokenABefore = corePoolKey.currency0.balanceOfSelf();
+        uint256 selfBalanceOfTokenABefore = corePoolKey.currency0.balanceOfSelf();
 
         swapRouter.swap(
             corePoolKey,
@@ -235,7 +193,7 @@ contract CSMMTest is Test, Deployers {
             ZERO_BYTES
         );
 
-        uint selfBalanceOfTokenAAfter = corePoolKey.currency0.balanceOfSelf();
+        uint256 selfBalanceOfTokenAAfter = corePoolKey.currency0.balanceOfSelf();
 
         assertEq(selfBalanceOfTokenABefore - selfBalanceOfTokenAAfter, 1e18);
     }
@@ -254,13 +212,11 @@ contract CSMMTest is Test, Deployers {
         );
 
         // add some liquidity to the core pool since it is where swaps will actually take place and not the proxy pool
-        PoolSwapTest.TestSettings memory settings = PoolSwapTest.TestSettings({
-            takeClaims: false,
-            settleUsingBurn: false
-        });
+        PoolSwapTest.TestSettings memory settings =
+            PoolSwapTest.TestSettings({takeClaims: false, settleUsingBurn: false});
 
-        uint selfBalanceOfTokenABefore = proxyPoolKey.currency0.balanceOfSelf();
-        uint selfBalanceOfTokenBBefore = proxyPoolKey.currency1.balanceOfSelf();
+        uint256 selfBalanceOfTokenABefore = proxyPoolKey.currency0.balanceOfSelf();
+        uint256 selfBalanceOfTokenBBefore = proxyPoolKey.currency1.balanceOfSelf();
 
         uint256 swapAmount = 100;
         swapRouter.swap(
@@ -274,13 +230,10 @@ contract CSMMTest is Test, Deployers {
             abi.encode(address(this))
         );
 
-        uint selfBalanceOfTokenAAfter = proxyPoolKey.currency0.balanceOfSelf();
-        uint selfBalanceOfTokenBAfter = proxyPoolKey.currency1.balanceOfSelf();
+        uint256 selfBalanceOfTokenAAfter = proxyPoolKey.currency0.balanceOfSelf();
+        uint256 selfBalanceOfTokenBAfter = proxyPoolKey.currency1.balanceOfSelf();
 
-        assertEq(
-            selfBalanceOfTokenABefore - selfBalanceOfTokenAAfter,
-            swapAmount
-        );
+        assertEq(selfBalanceOfTokenABefore - selfBalanceOfTokenAAfter, swapAmount);
         assert(selfBalanceOfTokenBAfter > selfBalanceOfTokenBBefore);
     }
 
@@ -298,13 +251,11 @@ contract CSMMTest is Test, Deployers {
         );
 
         // add some liquidity to the core pool since it is where swaps will actually take place and not the proxy pool
-        PoolSwapTest.TestSettings memory settings = PoolSwapTest.TestSettings({
-            takeClaims: false,
-            settleUsingBurn: false
-        });
+        PoolSwapTest.TestSettings memory settings =
+            PoolSwapTest.TestSettings({takeClaims: false, settleUsingBurn: false});
 
-        uint selfBalanceOfTokenABefore = proxyPoolKey.currency0.balanceOfSelf();
-        uint selfBalanceOfTokenBBefore = proxyPoolKey.currency1.balanceOfSelf();
+        uint256 selfBalanceOfTokenABefore = proxyPoolKey.currency0.balanceOfSelf();
+        uint256 selfBalanceOfTokenBBefore = proxyPoolKey.currency1.balanceOfSelf();
 
         uint256 swapAmount = 100;
         swapRouter.swap(
@@ -318,13 +269,10 @@ contract CSMMTest is Test, Deployers {
             abi.encode(address(this))
         );
 
-        uint selfBalanceOfTokenAAfter = proxyPoolKey.currency0.balanceOfSelf();
-        uint selfBalanceOfTokenBAfter = proxyPoolKey.currency1.balanceOfSelf();
+        uint256 selfBalanceOfTokenAAfter = proxyPoolKey.currency0.balanceOfSelf();
+        uint256 selfBalanceOfTokenBAfter = proxyPoolKey.currency1.balanceOfSelf();
 
-        assertEq(
-            selfBalanceOfTokenBBefore - selfBalanceOfTokenBAfter,
-            swapAmount
-        );
+        assertEq(selfBalanceOfTokenBBefore - selfBalanceOfTokenBAfter, swapAmount);
         assert(selfBalanceOfTokenAAfter > selfBalanceOfTokenABefore);
     }
 
@@ -342,13 +290,11 @@ contract CSMMTest is Test, Deployers {
         );
 
         // add some liquidity to the core pool since it is where swaps will actually take place and not the proxy pool
-        PoolSwapTest.TestSettings memory settings = PoolSwapTest.TestSettings({
-            takeClaims: false,
-            settleUsingBurn: false
-        });
+        PoolSwapTest.TestSettings memory settings =
+            PoolSwapTest.TestSettings({takeClaims: false, settleUsingBurn: false});
 
-        uint selfBalanceOfTokenABefore = proxyPoolKey.currency0.balanceOfSelf();
-        uint selfBalanceOfTokenBBefore = proxyPoolKey.currency1.balanceOfSelf();
+        uint256 selfBalanceOfTokenABefore = proxyPoolKey.currency0.balanceOfSelf();
+        uint256 selfBalanceOfTokenBBefore = proxyPoolKey.currency1.balanceOfSelf();
 
         uint256 swapAmount = 100;
         swapRouter.swap(
@@ -362,15 +308,12 @@ contract CSMMTest is Test, Deployers {
             abi.encode(address(this))
         );
 
-        uint selfBalanceOfTokenAAfter = proxyPoolKey.currency0.balanceOfSelf();
-        uint selfBalanceOfTokenBAfter = proxyPoolKey.currency1.balanceOfSelf();
+        uint256 selfBalanceOfTokenAAfter = proxyPoolKey.currency0.balanceOfSelf();
+        uint256 selfBalanceOfTokenBAfter = proxyPoolKey.currency1.balanceOfSelf();
 
         assert(selfBalanceOfTokenABefore > selfBalanceOfTokenAAfter);
 
-        assertEq(
-            selfBalanceOfTokenBAfter,
-            selfBalanceOfTokenBBefore + swapAmount
-        );
+        assertEq(selfBalanceOfTokenBAfter, selfBalanceOfTokenBBefore + swapAmount);
     }
 
     function test_swap_exactOutput_oneForZeroOnProxy() public {
@@ -387,13 +330,11 @@ contract CSMMTest is Test, Deployers {
         );
 
         // add some liquidity to the core pool since it is where swaps will actually take place and not the proxy pool
-        PoolSwapTest.TestSettings memory settings = PoolSwapTest.TestSettings({
-            takeClaims: false,
-            settleUsingBurn: false
-        });
+        PoolSwapTest.TestSettings memory settings =
+            PoolSwapTest.TestSettings({takeClaims: false, settleUsingBurn: false});
 
-        uint selfBalanceOfTokenABefore = proxyPoolKey.currency0.balanceOfSelf();
-        uint selfBalanceOfTokenBBefore = proxyPoolKey.currency1.balanceOfSelf();
+        uint256 selfBalanceOfTokenABefore = proxyPoolKey.currency0.balanceOfSelf();
+        uint256 selfBalanceOfTokenBBefore = proxyPoolKey.currency1.balanceOfSelf();
 
         uint256 swapAmount = 100;
         swapRouter.swap(
@@ -407,14 +348,11 @@ contract CSMMTest is Test, Deployers {
             abi.encode(address(this))
         );
 
-        uint selfBalanceOfTokenAAfter = proxyPoolKey.currency0.balanceOfSelf();
-        uint selfBalanceOfTokenBAfter = proxyPoolKey.currency1.balanceOfSelf();
+        uint256 selfBalanceOfTokenAAfter = proxyPoolKey.currency0.balanceOfSelf();
+        uint256 selfBalanceOfTokenBAfter = proxyPoolKey.currency1.balanceOfSelf();
 
         assert(selfBalanceOfTokenBBefore > selfBalanceOfTokenBAfter);
 
-        assertEq(
-            selfBalanceOfTokenAAfter,
-            selfBalanceOfTokenABefore + swapAmount
-        );
+        assertEq(selfBalanceOfTokenAAfter, selfBalanceOfTokenABefore + swapAmount);
     }
 }
